--- src/main/java/frc/robot/commands/Chassis/presets/GoToClosestPlaceToPlace.java
@@ -1,69 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-package frc.robot.commands.Chassis.presets;
-
-import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.Chassis;
-import frc.robot.supportingClasses.Auton.AutonManager;
-
-/** A command to go to whatever origin is */
-public class GoToClosestPlaceToPlace extends CommandBase {
-  private final Chassis m_chassis;
-  private final AutonManager autonManager;
-  private Thread m_thread;
-  private boolean firstHit = true;
-
-  private CommandBase autonCommand;
-
-  /**
-   * Creates a new goToOrigin Command.
-   *
-   * @param chassis The subsystem used by this command.
-   */
-  public GoToClosestPlaceToPlace(Chassis chassis, AutonManager manager) {
-    m_chassis = chassis;
-    // Use addRequirements() here to declare subsystem dependencies.
-    autonManager = manager;
-    addRequirements(chassis);
-  }
-
-  // Called when the command is initially scheduled.
-  @Override
-  public void initialize() {
-    firstHit = true;
-    autonCommand = null;
-
-    m_thread = new Thread(() -> autonCommand = autonManager.makeCmdToGoToPlace(m_chassis.getPose2d()));
-    m_thread.start();
-  }
-
-  // Called every time the scheduler runs while the command is scheduled.
-  @Override
-  public void execute() {
-    if (autonCommand != null) {
-      if (firstHit) {
-          autonCommand.initialize();
-          firstHit = false;
-      } else {
-        autonCommand.execute();
-      }
-    }
-  }
-
-  // Called once the command ends or is interrupted.
-  @Override
-  public void end(boolean interrupted) {
-    autonCommand.end(interrupted);
-  }
-
-  // Returns true when the command should end.
-  @Override
-  public boolean isFinished() {
-    if (autonCommand != null) {
-      return autonCommand.isFinished();
-    }
-    return false;
-  }
-}
