+++ src/main/java/frc/robot/commands/Chassis/presets/GoToClosestPlacementPosition.java
@@ -9,19 +9,19 @@ import frc.robot.supportingClasses.AutomaticallyGoToALocation;
 import frc.robot.supportingClasses.Auton.AutonManager;
 
 /**
- * A command to go to 0, 0 automatically
+ * A command to go to the closest placement position
  * Makes an object that sets the parent's ({@link AutomaticallyGoToALocation}) runnable.
  */
-public class GoToOrigin extends AutomaticallyGoToALocation {
+public class GoToClosestPlacementPosition extends AutomaticallyGoToALocation {
 
   /**
-   * Makes a new GoToOrigin object. calls the super's constructor to configure the runnable
+   * Makes a new GoToClosestPlacementPosition object. calls the super's constructor to configure the runnable
    *
    * @param m_chassis the chassis subsystem which gets required and used for the current position of the bot
    * @param manager the auton manager singleton which is used to generate the path as a runnable
    */
-  public GoToOrigin(Chassis m_chassis, AutonManager manager) {
-    super(m_chassis, () -> manager.backToStart(m_chassis.getPose2d()));
+  public GoToClosestPlacementPosition(Chassis m_chassis, AutonManager manager) {
+    super(m_chassis, () -> manager.makeCmdToGoToPlace(m_chassis.getPose2d()));
   }
 
 }
